feat: Split rank related stuff into dedicated RanksModule
