+++ server/src/modules/maps/maps.service.ts
@@ -3,14 +3,12 @@ import {
     BadRequestException,
     ConflictException,
     ForbiddenException,
-    ImATeapotException,
     Injectable,
     NotFoundException,
     StreamableFile
 } from '@nestjs/common';
-import { Map as MapDB, MapCredit, MapTrack, Prisma, Rank } from '@prisma/client';
+import { Map as MapDB, MapCredit, MapTrack, Prisma } from '@prisma/client';
 import { CreateMapDto, MapDto, UpdateMapDto } from '@common/dto/map/map.dto';
-import { RankDto } from '@common/dto/run/rank.dto';
 import { PaginatedResponseDto } from '@common/dto/paginated-response.dto';
 import { MapsRepoService } from '../repo/maps-repo.service';
 import { MapCreditType, MapStatus } from '@common/enums/map.enum';
@@ -21,18 +19,13 @@ import { ActivityType } from '@common/enums/activity.enum';
 import { CreateMapCreditDto, MapCreditDto, UpdateMapCreditDto } from '@common/dto/map/map-credit.dto';
 import { MapInfoDto, UpdateMapInfoDto } from '@common/dto/map/map-info.dto';
 import { MapTrackDto } from '@common/dto/map/map-track.dto';
-import {
-    AdminCtlMapsGetAllQuery,
-    MapsCtlGetAllQuery,
-    MapRanksGetQuery,
-    MapRankGetNumberQuery
-} from '@common/dto/query/map-queries.dto';
+import { AdminCtlMapsGetAllQuery, MapsCtlGetAllQuery } from '@common/dto/query/map-queries.dto';
 import { MapImageDto } from '@common/dto/map/map-image.dto';
 import { FileStoreCloudFile } from '../filestore/file-store.interface';
 import { ConfigService } from '@nestjs/config';
 import sharp from 'sharp';
 import { RunsService } from '../runs/runs.service';
-import { SteamService } from '@modules/steam/steam.service';
+import { RanksService } from '@modules/ranks/ranks.service';
 
 @Injectable()
 export class MapsService {
@@ -42,7 +35,7 @@ export class MapsService {
         private readonly fileCloudService: FileStoreCloudService,
         private readonly config: ConfigService,
         private readonly runsService: RunsService,
-        private readonly steamService: SteamService
+        private readonly ranksService: RanksService
     ) {}
 
     //#region Maps
@@ -662,156 +655,6 @@ export class MapsService {
         });
 
         return tracks.map((x) => DtoFactory(MapTrackDto, x));
-    }
-
-    //#endregion
-
-    // TODO: Move to own service!
-    //#region Ranks
-
-    async getRanks(mapID: number, query: MapRanksGetQuery): Promise<PaginatedResponseDto<RankDto>> {
-        const map = await this.mapRepo.get(mapID);
-
-        if (!map) throw new NotFoundException('Map not found');
-
-        const where: Prisma.RankWhereInput = {
-            mapID: mapID,
-            flags: query.flags
-        };
-
-        if (query.playerID) where.userID = query.playerID;
-        if (query.playerIDs) where.userID = { in: query.playerIDs };
-
-        const include = { run: true, user: true };
-
-        const order: Prisma.RankOrderByWithAggregationInput = {};
-        if (query.orderByDate !== undefined) order.createdAt = query.orderByDate ? 'desc' : 'asc';
-        else order.rank = 'asc';
-
-        const dbResponse = await this.mapRepo.getRanks(where, include, order, query.skip, query.take);
-
-        if (!dbResponse) throw new NotFoundException('No ranks found for map');
-
-        this.formatRanksDbResponse(dbResponse[0]);
-
-        return new PaginatedResponseDto(RankDto, dbResponse);
-    }
-
-    async getRankNumber(mapID: number, rankNumber: number, query: MapRankGetNumberQuery): Promise<RankDto> {
-        const map = await this.mapRepo.get(mapID);
-
-        if (!map) throw new NotFoundException('Map not found');
-
-        const where: Prisma.RankWhereInput = {
-            mapID: mapID,
-            rank: rankNumber,
-            flags: 0,
-            run: {
-                trackNum: 0,
-                zoneNum: 0
-            }
-        };
-
-        if (query.flags) where.flags = query.flags;
-        if (query.trackNum) where.run.trackNum = query.trackNum;
-        if (query.zoneNum) where.run.zoneNum = query.zoneNum;
-
-        const include = { run: true, user: true };
-
-        const dbResponse = (await this.mapRepo.getRank(where, include)) as any;
-
-        if (!dbResponse) throw new NotFoundException('Rank not found');
-
-        // Same approach as formatRanksDbResponse
-        dbResponse.trackNum = (dbResponse as any).run.trackNum;
-        dbResponse.zoneNum = (dbResponse as any).run.zoneNum;
-
-        return DtoFactory(RankDto, dbResponse);
-    }
-
-    async getRankAround(userID: number, mapID: number, query: MapRankGetNumberQuery): Promise<RankDto[]> {
-        const where: Prisma.RankWhereInput = {
-            mapID: mapID,
-            flags: 0,
-            userID: userID,
-            run: {
-                trackNum: 0,
-                zoneNum: 0
-            }
-        };
-
-        if (query.flags) where.flags = query.flags;
-        if (query.trackNum) where.run.trackNum = query.trackNum;
-        if (query.zoneNum) where.run.zoneNum = query.zoneNum;
-
-        const include = { run: true, user: true };
-
-        const order: Prisma.RankOrderByWithAggregationInput = { rank: 'asc' };
-
-        const userRankInfo = await this.mapRepo.getRank(where, include);
-
-        if (!userRankInfo) throw new NotFoundException('No personal best found');
-
-        const userRank = userRankInfo.rank;
-
-        // Reuse the previous query
-        where.userID = undefined;
-
-        // Don't care about the count
-        const [ranks] = await this.mapRepo.getRanks(
-            where,
-            include,
-            order,
-            // Minus 6 here because offset will skip the number of rows provided
-            // Example: if you want to offset to rank 9, you set offset to 8
-            Math.max(userRank - 6, 0),
-            11 // 5 + yours + 5
-        );
-
-        this.formatRanksDbResponse(ranks);
-
-        return ranks.map((rank) => DtoFactory(RankDto, rank));
-    }
-
-    async getRankFriends(steamID: bigint, mapID: number, query: MapRankGetNumberQuery): Promise<RankDto[]> {
-        const map = await this.mapRepo.get(mapID);
-
-        if (!map) throw new NotFoundException('Map not found');
-
-        const steamFriends = await this.steamService.getSteamFriends(steamID);
-
-        if (steamFriends.length === 0) throw new ImATeapotException('No friends detected :(');
-
-        const friendSteamIDs = steamFriends.map((item) => BigInt(item.steamid));
-
-        const where: Prisma.RankWhereInput = {
-            mapID: mapID,
-            flags: 0,
-            user: { steamID: { in: friendSteamIDs } },
-            run: { trackNum: 0, zoneNum: 0 }
-        };
-
-        if (query.flags) where.flags = query.flags;
-        if (query.trackNum) where.run.trackNum = query.trackNum;
-        if (query.zoneNum) where.run.zoneNum = query.zoneNum;
-
-        const include = { run: true, user: true };
-
-        // Don't care about the count
-        const [ranks] = await this.mapRepo.getRanks(where, include);
-
-        this.formatRanksDbResponse(ranks);
-
-        return ranks.map((rank) => DtoFactory(RankDto, rank));
-    }
-
-    // This is done because the MapRankDto still contains trackNum and zoneNum to conform to old API
-    // but Rank model doesn't. Probably worth changing frontend/game code in future.
-    private formatRanksDbResponse(ranks: (Rank & { trackNum?: any; zoneNum?: any })[]) {
-        for (const mapRank of ranks) {
-            mapRank.trackNum = (mapRank as any).run.trackNum;
-            mapRank.zoneNum = (mapRank as any).run.zoneNum;
-        }
     }
 
     //#endregion
