+++ server/src/modules/ranks/ranks.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { RanksService } from './ranks.service';
+import { RepoModule } from '@modules/repo/repo.module';
+import { SteamModule } from '@modules/steam/steam.module';
+
+@Module({
+    imports: [RepoModule, SteamModule],
+    providers: [RanksService],
+    exports: [RanksService]
+})
+export class RanksModule {}
